fix : 카카오톡 미리보기 안되는 버그 해결 (CC-153) (#40)

+++ src/main/java/ai/softeer/caecae/shorturl/domain/request/ShareUrlRequestDto.java
@@ -2,6 +2,6 @@ package ai.softeer.caecae.shorturl.domain.request;
 
 public record ShareUrlRequestDto(
         double distance,
-        int percentage
+        double percentage
 ) {
 }
